# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FeedKeeper native API
  version: "1"
  license: { name: MIT, url: "https://github.com/visualfusion/feedkeeper/blob/main/LICENSE" }
  description: |
    The API for native apps, served under `/api/v1`. Design notes and what is still planned
    are in `docs/design/native-api.md`. A test keeps this file in step with the implemented routes.

    Authenticate with `Authorization: Bearer <token>`: a device token from pairing, or a personal
    access token. Read-only tokens can call every `GET` and nothing that changes data.
servers:
  - url: /api/v1
security:
  - bearer: []
...
paths:
...
  /discover:
    post:
      tags: [Subscriptions]
      summary: Find the feeds of a website
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [url], properties: { url: { type: string } } } } }
      responses:
        "200": { description: OK }
        "400": { $ref: "#/components/responses/Invalid" }
  /opml:
    get:
      tags: [Subscriptions]
      summary: Export subscriptions as OPML
      responses:
        "200": { description: OPML 2.0, content: { text/x-opml: { schema: { type: string } } } }
    post:
      tags: [Subscriptions]
      summary: Import subscriptions from OPML
      description: "Send the XML as the body (text/xml) or as JSON { opml: \"...\" }."
      responses:
        "200": { description: Import result }
        "400": { $ref: "#/components/responses/Invalid" }
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer }
...
  responses:
...
    Invalid: { description: "Invalid input: { error: invalid_input }" }
...